requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
crewai
crewai-tools
PyMuPDF
langchain-ollama
langchain-openai
arxiv
python-multipart
requests